requirements.txt
bcrypt==4.2.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
psycopg2==2.9.9
pydantic==2.8.2
SQLAlchemy==2.0.32
typing_extensions==4.12.2
requests==2.32.3
python-dotenv==1.0.1